backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
google-generativeai
python-dotenv
bcrypt
